Consolidate 4 GitHub workflows into one ci.yml and inline the Svelte config
Replaces test.yml, lint.yml, gh-pages.yml and link-check.yml with a single ci.yml.
Moves the SvelteKit config inline into vite.config.ts and drops svelte.config.ts,
updating the CodeExample docs and readme badges to match.
