--- .github/workflows/gh-pages.yml
@@ -1,42 +1,0 @@
-name: GitHub Pages
-
-on:
-  push:
-    branches: [main]
-  pull_request:
-    branches: [main]
-  workflow_dispatch:
-
-permissions:
-  pages: write
-  id-token: write
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v6
-
-      - uses: actions/setup-node@v6
-        with:
-          node-version: 24
-
-      - run: npm install
-      - run: npm run build:site
-        env:
-          BASE_PATH: /${{ github.event.repository.name }}
-
-      - uses: actions/upload-pages-artifact@v4
-        with:
-          path: build
-
-  deploy:
-    if: github.ref == 'refs/heads/main'
-    needs: build
-    runs-on: ubuntu-latest
-    environment:
-      name: github-pages
-      url: ${{ steps.deploy.outputs.page_url }}
-    steps:
-      - id: deploy
-        uses: actions/deploy-pages@v4
