--- .github/workflows/link-check.yml
@@ -1,21 +1,0 @@
-name: Link check
-
-on:
-  push:
-    branches: [main]
-  pull_request:
-    branches: [main]
-
-jobs:
-  link-check:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Check out repo
-        uses: actions/checkout@v6
-
-      - name: Discover broken links
-        uses: lycheeverse/lychee-action@v2
-        with:
-          # Deployed docs and unreleased compare tags are unavailable until after merge.
-          # The prerender crawl validates internal docs links.
-          args: --exclude '%7B' --exclude '^file://' --exclude '^https://svelte-widgets\.janosh\.dev' --exclude '^https://github\.com/janosh/svelte-widgets/compare/' --root-dir . --accept 100..=103,200..=299,403,429,502,503 -- ./**/*.{md,svelte,ts}
