--- .github/workflows/lint.yml
@@ -1,25 +1,0 @@
-name: Lint
-
-on:
-  push:
-    branches: [main]
-  pull_request:
-    branches: [main]
-
-jobs:
-  check:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v6
-
-      - uses: actions/setup-node@v6
-        with:
-          node-version: 24
-
-      - run: npm install
-
-      - name: Lint, format, and type check
-        run: npx vp check
-
-      - name: Svelte type check
-        run: npx svelte-kit sync && npx svelte-check-rs --threshold error
