--- .github/workflows/test.yml
@@ -1,43 +1,0 @@
-name: Tests
-
-on:
-  pull_request:
-    branches: [main]
-  push:
-    branches: [main]
-  workflow_dispatch:
-
-jobs:
-  unit:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v6
-
-      - uses: actions/setup-node@v6
-        with:
-          node-version: 24
-
-      - run: npm install
-
-      - name: Unit tests with coverage
-        run: npm run test:coverage
-
-      - name: Package smoke test
-        run: npm run test:package
-
-  e2e:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v6
-
-      - uses: actions/setup-node@v6
-        with:
-          node-version: 24
-
-      - run: npm install
-
-      - name: Install Playwright
-        run: npx playwright install chromium
-
-      - name: E2E tests
-        run: npx playwright test
