+++ vite.config.ts
@@ -1,10 +1,77 @@
+import adapter from '@sveltejs/adapter-static'
 import { sveltekit } from '@sveltejs/kit/vite'
+import { mdsvex } from 'mdsvex'
+import pkg from './package.json' with { type: 'json' }
 import { generate_icons } from './scripts/generate-icons.ts'
+import { heading_ids } from './src/lib/heading-anchors.ts'
+import { katex_preprocess } from './src/lib/katex.ts'
+import {
+  mdsvex_transform,
+  starry_night_highlighter,
+} from './src/lib/live-examples/index.ts'
 import live_examples from './src/lib/live-examples/vite-plugin.ts'
 import source_links from './src/lib/source-links/vite-plugin.ts'
 import { make_config } from './src/lib/vite-config.ts'
 
 await generate_icons()
+
+const base_segment = (process.env.BASE_PATH ?? ``).replaceAll(/^\/+|\/+$/gu, ``)
+const base_path: `` | `/${string}` = base_segment ? `/${base_segment}` : ``
+const remark_plugins = [
+  [
+    mdsvex_transform,
+    {
+      defaults: {
+        Wrapper: `/src/lib/CodeExample.svelte`,
+        repo: pkg.repository,
+        collapsible: true,
+        hide_style: true,
+      },
+    },
+  ],
+]
+const { before: katex_before, after: katex_after } = katex_preprocess()
+
+// passed inline to sveltekit() (Kit >= 2.62) so no separate svelte.config.ts is needed;
+// kit options (adapter, alias, paths, prerender) sit at the top level rather than under `kit`.
+// svelte-package only reads svelte.config.*, so it packages src/lib with default config: nothing
+// in src/lib relies on these preprocessors or aliases and the `package` script drops the one .md.
+const svelte_config = {
+  extensions: [`.svelte`, `.md`],
+
+  // KaTeX before/after mdsvex so markdown never sees rendered HTML; heading IDs last.
+  preprocess: [
+    katex_before,
+    mdsvex({
+      remarkPlugins: remark_plugins,
+      extensions: [`.md`],
+      highlight: { highlighter: starry_night_highlighter },
+    }),
+    katex_after,
+    heading_ids(),
+  ],
+
+  adapter: adapter(),
+  paths: { base: base_path },
+
+  alias: {
+    $root: `.`,
+    $site: `./src/site`,
+    'svelte-widgets': `./src/lib`,
+  },
+
+  prerender: {
+    handleHttpError: ({ status, referrer, message }) => {
+      // Ignore 404s from the /nav demo page which contains links to non-existent routes
+      if (status === 404 && referrer === `${base_path}/nav`) return
+      throw new Error(message)
+    },
+  },
+
+  vitePlugin: {
+    inspector: true,
+  },
+} satisfies Parameters<typeof sveltekit>[0]
 
 export default {
   // shared lint/fmt/build/staged, published as svelte-widgets/vite-config
@@ -18,7 +85,7 @@ export default {
     },
   }),
 
-  plugins: [sveltekit(), ...live_examples(), source_links()],
+  plugins: [sveltekit(svelte_config), ...live_examples(), source_links()],
 
   test: {
     include: [`tests/vitest/**/*.test.ts`],
